Divide the API structure into types.go
Narrow down the structure to the minimum necessary members.
