+++ pkg/client/client.go
@@ -44,128 +44,6 @@ type CDIClient struct {
 	TokenSource oauth2.TokenSource
 }
 
-type IMToken struct {
-	AccessToken      string `json:"access_token"`
-	ExpiresIn        int64  `json:"expires_in"`
-	RefreshExpiresIn int64  `json:"refresh_expires_in"`
-	RefreshToken     string `json:"refresh_token"`
-	TokenType        string `json:"token_type"`
-	IDToken          string `json:"id_token"`
-	NotBeforePolicy  int64  `json:"not-before-policy"`
-	SessionState     string `json:"session_state"`
-	Scope            string `json:"scope"`
-}
-
-type FMMachineList struct {
-	Data FMMachines `json:"data"`
-}
-
-type FMMachines struct {
-	Machines []FMMachine `json:"machines"`
-}
-
-type FMMachine struct {
-	FabricUUID   *string             `json:"fabric_uuid"`
-	FabricID     *int                `json:"fabric_id"`
-	MachineUUID  string              `json:"mach_uuid"`
-	MachineID    int                 `json:"mach_id"`
-	MachineName  string              `json:"mach_name"`
-	MachineOwner string              `json:"mach_owner"`
-	Resources    []FMMachineResource `json:"resources"`
-}
-
-type FMMachineResource struct {
-	ResourceUUID     string `json:"res_uuid"`
-	ResourceName     string `json:"res_name"`
-	ResourceType     string `json:"res_type"`
-	ResourceStatus   int    `json:"res_status"`
-	ResourceOPStatus string `json:"res_op_status"`
-}
-
-type FMAvailableReservedResources struct {
-	FabricUUID          string `json:"fabric_uuid"`
-	FabricID            int    `json:"fabric_id"`
-	ReservedResourceNum int    `json:"reserved_res_num_per_fabric"`
-}
-
-type Condition struct {
-	Column   string `json:"column"`
-	Operator string `json:"operator"`
-	Value    string `json:"value"`
-}
-
-type CMNodeGroups struct {
-	NodeGroups []CMNodeGroup `json:"nodegroups"`
-}
-
-type CMNodeGroup struct {
-	UUID      string `json:"uuid"`
-	Name      string `json:"name"`
-	NodeCount int    `json:"node_count"`
-	Role      string `json:"role"`
-}
-
-type CMNodeGroupInfo struct {
-	UUID         string                `json:"uuid"`
-	Name         string                `json:"name"`
-	Composable   bool                  `json:"composable"`
-	NodeCount    int                   `json:"node_count"`
-	Role         string                `json:"role"`
-	MinNodeCount int                   `json:"min_node_count"`
-	MaxNodeCount int                   `json:"max_node_count"`
-	Status       string                `json:"status"`
-	StatusReason string                `json:"status_reason"`
-	Resources    []CMNodeGroupResource `json:"resources"`
-	NodeIDs      []string              `json:"node_ids"`
-	MachineIDs   []string              `json:"mach_ids"`
-}
-
-type CMNodeGroupResource struct {
-	ResourceName     string `json:"resource_name"`
-	ResourceType     string `json:"resource_type"`
-	ModelName        string `json:"model_name"`
-	MinResourceCount *int   `json:"min_resource_count"`
-	MaxResourceCount *int   `json:"max_resource_count"`
-}
-
-type CMNodeDetails struct {
-	Data CMTenant `json:"data"`
-}
-
-type CMTenant struct {
-	TenantUUID string    `json:"tenant_uuid"`
-	Cluster    CMCluster `json:"cluster"`
-}
-
-type CMCluster struct {
-	ClusterUUID string    `json:"cluster_uuid"`
-	Machine     CMMachine `json:"machine"`
-}
-
-type CMMachine struct {
-	UUID     string      `json:"uuid"`
-	Name     string      `json:"name"`
-	ResSpecs []CMResSpec `json:"resspecs"`
-}
-
-type CMResSpec struct {
-	SpecUUID        string     `json:"spec_uuid"`
-	Type            string     `json:"type"`
-	Selector        CMSelector `json:"selector"`
-	MinResSpecCount *int       `json:"min_resspec_count"`
-	MaxResSpecCount *int       `json:"max_resspec_count"`
-	DeviceCount     int        `json:"device_count"`
-}
-
-type CMSelector struct {
-	Version    string       `json:"version"`
-	Expression CMExpression `json:"expression"`
-}
-
-type CMExpression struct {
-	Conditions []Condition `json:"conditions"`
-}
-
 type RequestIDKey struct{}
 
 func BuildCDIClient(config *config.Config, kc *kube_utils.KubeControllers) (*CDIClient, error) {
