+++ pkg/client/client_test.go
@@ -81,15 +81,8 @@ func TestCDIClientGetIMToken(t *testing.T) {
 			password: "pass",
 			realm:    "CDI_DRA_Test",
 			expectedToken: &IMToken{
-				AccessToken:      "token1" + "." + base64.RawURLEncoding.EncodeToString([]byte(`{"exp":2069550000}`)),
-				ExpiresIn:        1,
-				RefreshExpiresIn: 2,
-				RefreshToken:     "token2",
-				TokenType:        "Bearer",
-				IDToken:          "token3",
-				NotBeforePolicy:  3,
-				SessionState:     "efffca5t4",
-				Scope:            "test profile",
+				AccessToken: "token1" + "." + base64.RawURLEncoding.EncodeToString([]byte(`{"exp":2069550000}`)),
+				ExpiresIn:   1,
 			},
 			expectedErr: false,
 		},
